@@ -6,8 +6,8 @@ repository: https://github.com/chulwoo-park/timelines/
 issue_tracker: https://github.com/chulwoo-park/timelines/issues
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: '>=1.17.0'
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
@@ -18,4 +18,4 @@ dev_dependencies:
     sdk: flutter
   pedantic: ^1.11.0
 
-flutter: 
+flutter:
