@@ -74,7 +74,7 @@ class _ThemePageState extends State<ThemePage> {
                     children: [
                       Text(
                         'TimelineTheme',
-                        style: Theme.of(context).textTheme.headline6,
+                        style: Theme.of(context).textTheme.headlineSmall,
                       ),
                       _ThemeDropdown(
                         title: 'Direction',
@@ -140,7 +140,7 @@ class _ThemePageState extends State<ThemePage> {
                     children: [
                       Text(
                         'IndicatorTheme',
-                        style: Theme.of(context).textTheme.headline6,
+                        style: Theme.of(context).textTheme.headlineSmall,
                       ),
                       _ThemeDropdown<Color?>(
                         title: 'Color',
@@ -192,7 +192,7 @@ class _ThemePageState extends State<ThemePage> {
                     children: [
                       Text(
                         'ConnectorTheme',
-                        style: Theme.of(context).textTheme.headline6,
+                        style: Theme.of(context).textTheme.headlineSmall,
                       ),
                       _ThemeDropdown<Color?>(
                         title: 'Color',
